moved user controllerr
